test(bmr): consolidate legacy ownership coverage
Move scattered BMR wrapper and legacy tests into the canonical premium BMR suite without changing the covered runtime contract or test count. Remove obsolete dedicated test files and make the metrics-aware TestClient seam type-safe for the consolidated suite.
