+++ tests/test_premium_bmr_api.py
@@ -8,14 +8,24 @@ Tests cover:
 - Premium feature integration
 """
 
+import asyncio
 import os
+import sys
+import types
+from types import ModuleType
+from typing import Any, cast
 from unittest.mock import patch
-from typing import cast
 
 import pytest
+from fastapi import HTTPException
 from fastapi.testclient import TestClient
+from starlette.types import ASGIApp
 
-from app import app
+import app as app_package
+from app import app as fastapi_app
+from app.utils import nutrition_wrappers
+import legacy_app
+from tests._client import get_client
 
 
 def test_bmr_rejects_invalid_sex() -> None:
@@ -31,7 +41,7 @@ def test_bmr_rejects_invalid_sex() -> None:
 
 @pytest.fixture
 def client() -> TestClient:
-    return TestClient(app)
+    return TestClient(fastapi_app)
 
 
 class TestPremiumBMRAPI:
@@ -369,6 +379,975 @@ class TestPremiumBMRAPI:
         data = response.json()
         assert all(bmr > 0 for bmr in data["bmr"].values())
         assert all(tdee > bmr for bmr, tdee in zip(data["bmr"].values(), data["tdee"].values()))
+
+
+def test_resolve_prefers_app_over_app_module(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Test that app.<name> is preferred over app.app_module.<name>."""
+
+    def f_app(*a: object, **k: object) -> dict[str, float]:
+        return {"mifflin": 1.0}
+
+    def f_appmod(*a: object, **k: object) -> dict[str, float]:
+        return {"mifflin": 2.0}
+
+    fake_appmod = types.SimpleNamespace(calculate_all_bmr=f_appmod)
+    fake_app = types.SimpleNamespace(calculate_all_bmr=f_app, app_module=fake_appmod)
+    fake_alias = types.SimpleNamespace(calculate_all_bmr=lambda *a, **k: {"mifflin": 3.0})
+
+    # Patch the seam function instead of sys.modules directly
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (fake_app, fake_alias, fake_appmod),
+        raising=True,
+    )
+
+    fn = nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+    assert fn is f_app
+
+
+def test_resolve_falls_back_to_app_module(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Test that app.app_module.<name> is used when app.<name> is not available."""
+
+    def f_appmod(*a: object, **k: object) -> dict[str, float]:
+        return {"mifflin": 2.0}
+
+    fake_appmod = types.SimpleNamespace(calculate_all_bmr=f_appmod)
+    fake_app = types.SimpleNamespace(app_module=fake_appmod)  # No calculate_all_bmr on app
+
+    # Patch the seam function instead of sys.modules directly
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (fake_app, None, fake_appmod),
+        raising=True,
+    )
+
+    fn = nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+    assert fn is f_appmod
+
+
+def test_resolve_falls_back_to_sys_modules_app_module(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Test that sys.modules['app_module'].<name> is used when app/app.app_module are not available."""
+
+    def f_alias(*a: object, **k: object) -> dict[str, float]:
+        return {"mifflin": 3.0}
+
+    fake_app = types.SimpleNamespace()  # No calculate_all_bmr, no app_module
+    fake_appmod = types.SimpleNamespace(calculate_all_bmr=f_alias)
+
+    # Patch the seam function instead of sys.modules directly
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (fake_app, fake_appmod, None),
+        raising=True,
+    )
+
+    fn = nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+    assert fn is f_alias
+
+
+def test_resolve_unknown_name_raises() -> None:
+    """Test that unknown callable name raises ImportError."""
+    with pytest.raises(ImportError, match="unknown nutrition callable"):
+        nutrition_wrappers._resolve_nutrition_callable("nope")
+
+
+def test_resolve_falls_back_to_import_seam(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Test that nutrition_core import seam is used when all other paths fail."""
+
+    def seam_fn(*a: object, **k: object) -> dict[str, float]:
+        return {"mifflin": 123.0}
+
+    # Patch the seam function to return None modules (all paths fail)
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (None, None, None),
+        raising=True,
+    )
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_bmr",
+        lambda: seam_fn,
+        raising=True,
+    )
+
+    fn = nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+    assert fn is seam_fn
+    assert fn() == {"mifflin": 123.0}
+
+
+def test_resolve_import_seam_none_raises(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Test that ImportError is raised when import seam returns None."""
+    # Patch the seam function to return None modules (all paths fail)
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (None, None, None),
+        raising=True,
+    )
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_bmr",
+        lambda: None,
+        raising=True,
+    )
+
+    with pytest.raises(ImportError, match="not available"):
+        nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+
+
+def test_resolve_tdee_uses_seam(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Test that TDEE resolution also uses import seams."""
+
+    def seam_fn(*a: object, **k: object) -> dict[str, int | float]:
+        return {"mifflin": 456.0}
+
+    # Patch the seam function to return None modules (all paths fail)
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (None, None, None),
+        raising=True,
+    )
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_tdee",
+        lambda: seam_fn,
+        raising=True,
+    )
+
+    fn = nutrition_wrappers._resolve_nutrition_callable("calculate_all_tdee")
+    assert fn is seam_fn
+    assert fn({"mifflin": 1500.0}, "moderate") == {"mifflin": 456.0}
+
+
+def test_resolve_skips_non_callable_attr_and_falls_back_to_seam(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Test that non-callable attributes are skipped and fallback to seam is used."""
+    # pkg exposes the "right" name but it's NOT callable
+    fake_pkg = types.SimpleNamespace(calculate_all_bmr="not-a-function")
+    fake_alias = None
+    fake_pkg_appmod = None
+
+    def seam_fn(*a: object, **k: object) -> dict[str, float]:
+        return {"mifflin": 123.0}
+
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_get_candidate_modules",
+        lambda: (fake_pkg, fake_alias, fake_pkg_appmod),
+        raising=True,
+    )
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_bmr",
+        lambda: seam_fn,
+        raising=True,
+    )
+
+    fn = nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+    assert fn is seam_fn
+    assert fn() == {"mifflin": 123.0}
+
+
+def test_import_nutrition_core_import_seams(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Cover _import_nutrition_core_* implementations (ImportError + happy paths)."""
+
+    def _boom(_name: str) -> types.ModuleType:
+        raise ImportError("boom")
+
+    monkeypatch.setattr(nutrition_wrappers.importlib, "import_module", _boom)
+    assert nutrition_wrappers._import_nutrition_core_bmr() is None
+    assert nutrition_wrappers._import_nutrition_core_tdee() is None
+
+    mod = types.ModuleType("nutrition_core")
+
+    def calculate_all_bmr(*_args: Any, **_kwargs: Any) -> dict[str, float]:
+        return {"mifflin": 1500.0}
+
+    def calculate_all_tdee(*_args: Any, **_kwargs: Any) -> dict[str, int | float]:
+        return {"mifflin": 2000.0}
+
+    setattr(mod, "calculate_all_bmr", calculate_all_bmr)
+    setattr(mod, "calculate_all_tdee", calculate_all_tdee)
+
+    def _import_module(name: str) -> types.ModuleType:
+        assert name == "nutrition_core"
+        return mod
+
+    monkeypatch.setattr(nutrition_wrappers.importlib, "import_module", _import_module)
+    assert nutrition_wrappers._import_nutrition_core_bmr() is calculate_all_bmr
+    assert nutrition_wrappers._import_nutrition_core_tdee() is calculate_all_tdee
+
+
+def test_resolve_nutrition_callable_prefers_app_app_module_over_alias(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Cover resolution order: app.app_module wins over sys.modules['app_module']."""
+    pkg = types.ModuleType("app")
+    pkg_appmod = types.ModuleType("app.app_module")
+    alias_pkg = types.ModuleType("app_module")
+
+    def appmod_bmr(*_args: Any, **_kwargs: Any) -> dict[str, float]:
+        return {"mifflin": 1500.0}
+
+    def alias_bmr(*_args: Any, **_kwargs: Any) -> dict[str, float]:
+        return {"mifflin": 1400.0}
+
+    setattr(pkg_appmod, "calculate_all_bmr", appmod_bmr)
+    setattr(alias_pkg, "calculate_all_bmr", alias_bmr)
+    setattr(pkg, "app_module", pkg_appmod)
+
+    monkeypatch.setitem(sys.modules, "app", pkg)
+    monkeypatch.setitem(sys.modules, "app_module", alias_pkg)
+
+    resolved = nutrition_wrappers._resolve_nutrition_callable("calculate_all_bmr")
+    assert resolved is appmod_bmr
+
+
+def test_resolve_nutrition_callable_unknown_name_raises(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Cover unknown callable name branch."""
+    monkeypatch.setitem(sys.modules, "app", types.ModuleType("app"))
+    monkeypatch.setitem(sys.modules, "app_module", types.ModuleType("app_module"))
+
+    with pytest.raises(ImportError, match="unknown nutrition callable"):
+        nutrition_wrappers._resolve_nutrition_callable("unknown")
+
+
+def test_calculate_wrappers_fallback_to_nutrition_core_real_import_seams(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Cover wrapper fallback to nutrition_core via real import seams."""
+    mod = types.ModuleType("nutrition_core")
+
+    def calculate_all_bmr(
+        weight_kg: float,
+        height_cm: float,
+        age: int,
+        sex: str,
+        bodyfat: float | None,
+    ) -> dict[str, float]:
+        assert (weight_kg, height_cm, age, sex, bodyfat) == (70.0, 175.0, 30, "male", None)
+        return {"mifflin": 1500.0}
+
+    def calculate_all_tdee(bmr_results: dict[str, float], activity: str) -> dict[str, int | float]:
+        assert bmr_results == {"mifflin": 1500.0}
+        assert activity == "moderate"
+        return {"mifflin": 2000.0}
+
+    setattr(mod, "calculate_all_bmr", calculate_all_bmr)
+    setattr(mod, "calculate_all_tdee", calculate_all_tdee)
+
+    def _import_module(name: str) -> types.ModuleType:
+        assert name == "nutrition_core"
+        return mod
+
+    # Ensure wrappers don't resolve from app/app_module so fallback is exercised
+    monkeypatch.setitem(sys.modules, "app", types.ModuleType("app"))
+    monkeypatch.delitem(sys.modules, "app_module", raising=False)
+    monkeypatch.setattr(nutrition_wrappers.importlib, "import_module", _import_module)
+
+    bmr = nutrition_wrappers._calculate_all_bmr_wrapper(70.0, 175.0, 30, "male", bodyfat=None)
+    assert bmr == {"mifflin": 1500.0}
+
+    tdee = nutrition_wrappers._calculate_all_tdee_wrapper({"mifflin": 1500.0}, "moderate")
+    assert tdee == {"mifflin": 2000.0}
+
+
+def test_calculate_wrappers_import_error(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Ensure wrappers raise ImportError when their dependencies are missing."""
+
+    # Null out all visible locations so wrappers raise ImportError deterministically
+    for module in (
+        app_package,
+        getattr(app_package, "app_module", None),
+        sys.modules.get("app_module"),
+    ):
+        if module is not None:
+            monkeypatch.setattr(module, "calculate_all_bmr", None, raising=False)
+            monkeypatch.setattr(module, "calculate_all_tdee", None, raising=False)
+
+    # Block nutrition_core import seams (wrapper's fallback) by patching import functions
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_bmr",
+        lambda: None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_tdee",
+        lambda: None,
+        raising=False,
+    )
+
+    with pytest.raises(ImportError):
+        nutrition_wrappers._calculate_all_bmr_wrapper(70, 175, 30, "male")
+
+    with pytest.raises(ImportError):
+        nutrition_wrappers._calculate_all_tdee_wrapper({"mifflin": 1500}, "moderate")
+
+
+def test_calculate_all_bmr_wrapper_happy_path_nutrition_core(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Test BMR wrapper happy path when nutrition_core is available."""
+    calls: dict[str, object] = {}
+
+    def fake_bmr(*args: Any, **kwargs: Any) -> dict[str, float]:
+        calls["args"] = args
+        calls["kwargs"] = kwargs
+        return {"mifflin": 1500.0, "harris": 1600.0}
+
+    # Null out app/app_module paths to force nutrition_core fallback
+    for module in (
+        app_package,
+        getattr(app_package, "app_module", None),
+        sys.modules.get("app_module"),
+    ):
+        if module is not None:
+            monkeypatch.setattr(module, "calculate_all_bmr", None, raising=False)
+
+    # Patch import seam to return fake function
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_bmr",
+        lambda: fake_bmr,
+        raising=False,
+    )
+
+    res = nutrition_wrappers._calculate_all_bmr_wrapper(70.0, 175.0, 30, "male", bodyfat=None)
+    assert res == {"mifflin": 1500.0, "harris": 1600.0}
+    assert calls["args"] == (70.0, 175.0, 30, "male", None)
+    assert calls["kwargs"] == {}
+
+
+def test_calculate_all_tdee_wrapper_happy_path_nutrition_core(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    """Test TDEE wrapper happy path when nutrition_core is available."""
+    calls: dict[str, object] = {}
+
+    def fake_tdee(*args: Any, **kwargs: Any) -> dict[str, int | float]:
+        calls["args"] = args
+        calls["kwargs"] = kwargs
+        return {"mifflin": 2000.0, "harris": 2100.0}
+
+    # Null out app/app_module paths to force nutrition_core fallback
+    for module in (
+        app_package,
+        getattr(app_package, "app_module", None),
+        sys.modules.get("app_module"),
+    ):
+        if module is not None:
+            monkeypatch.setattr(module, "calculate_all_tdee", None, raising=False)
+
+    # Patch import seam to return fake function
+    monkeypatch.setattr(
+        nutrition_wrappers,
+        "_import_nutrition_core_tdee",
+        lambda: fake_tdee,
+        raising=False,
+    )
+
+    res = nutrition_wrappers._calculate_all_tdee_wrapper({"mifflin": 1500.0}, "moderate")
+    assert res == {"mifflin": 2000.0, "harris": 2100.0}
+    assert calls["args"] == ({"mifflin": 1500.0}, "moderate")
+    assert calls["kwargs"] == {}
+
+
+def _get_legacy_bmr_app() -> ASGIApp:
+    """Safely get the FastAPI app instance from main.py."""
+    import main
+
+    if getattr(main, "app", None) is None:
+        raise RuntimeError("FastAPI app in main.py is not initialized")
+    return cast(ASGIApp, main.app)
+
+
+legacy_bmr_client = TestClient(_get_legacy_bmr_app())
+
+
+class TestLegacyPremiumBMRComprehensive:
+    """Legacy BMR coverage formerly owned by the broad comprehensive suite."""
+
+    def test_premium_bmr_endpoint_module_not_available(self):
+        """Test /premium_bmr when nutrition module not available (lines 1180-1189)"""
+        with patch(
+            "app._calculate_all_bmr_wrapper",
+            side_effect=ImportError("nutrition_core module not available"),
+        ):
+            response = legacy_bmr_client.post(
+                "/premium_bmr",
+                json={
+                    "weight_kg": 70,
+                    "height_cm": 170,
+                    "age": 30,
+                    "sex": "male",
+                    "activity": "moderate",
+                    "lang": "en",
+                },
+            )
+            assert response.status_code == 503
+            assert "not available" in response.json()["detail"]
+
+    @patch("app._calculate_all_bmr_wrapper")
+    def test_premium_bmr_endpoint_value_error(self, mock_bmr):
+        """Test /premium_bmr with ValueError (lines 1235-1236)"""
+        mock_bmr.side_effect = ValueError("Invalid input data")
+
+        response = legacy_bmr_client.post(
+            "/premium_bmr",
+            json={
+                "weight_kg": -10,  # Invalid weight
+                "height_cm": 170,
+                "age": 30,
+                "sex": "male",
+                "activity": "moderate",
+                "lang": "en",
+            },
+        )
+        assert response.status_code == 400
+        assert "Invalid input" in response.json()["detail"]
+
+    @patch("app._calculate_all_bmr_wrapper")
+    def test_premium_bmr_endpoint_general_error(self, mock_bmr):
+        """Test /premium_bmr with general exception (lines 1237-1238)"""
+        mock_bmr.side_effect = Exception("Calculation failed")
+
+        response = legacy_bmr_client.post(
+            "/premium_bmr",
+            json={
+                "weight_kg": 70,
+                "height_cm": 170,
+                "age": 30,
+                "sex": "male",
+                "activity": "moderate",
+                "lang": "en",
+            },
+        )
+        assert response.status_code == 500
+        assert "BMR calculation failed" in response.json()["detail"]
+
+    def test_activity_level_descriptions(self):
+        """Test activity level descriptions in premium_bmr"""
+        with patch("app.calculate_all_bmr", return_value={"mifflin": 1800}):
+            with patch("app.calculate_all_tdee", return_value={"mifflin": 2200}):
+                response = legacy_bmr_client.post(
+                    "/premium_bmr",
+                    json={
+                        "weight_kg": 70,
+                        "height_cm": 170,
+                        "age": 30,
+                        "sex": "male",
+                        "activity": "very_active",
+                        "lang": "en",
+                    },
+                )
+                assert response.status_code == 200
+                data = response.json()
+                assert "activity_level" in data
+
+    def test_katch_bmr_note(self):
+        """Test Katch BMR formula note when bodyfat provided"""
+        with patch(
+            "app.calculate_all_bmr",
+            return_value={"mifflin": 1800, "katch": 1900},
+        ):
+            with patch(
+                "app.calculate_all_tdee",
+                return_value={"mifflin": 2200, "katch": 2300},
+            ):
+                response = legacy_bmr_client.post(
+                    "/premium_bmr",
+                    json={
+                        "weight_kg": 70,
+                        "height_cm": 170,
+                        "age": 30,
+                        "sex": "male",
+                        "activity": "moderate",
+                        "bodyfat": 15,
+                        "lang": "en",
+                    },
+                )
+                assert response.status_code == 200
+                data = response.json()
+                assert "notes" in data
+                # Should include Katch formula note when bodyfat is provided
+
+
+class TestLegacyPremiumBMRRuntimeFallback:
+    """Legacy BMR fallback coverage formerly owned by the extended suite."""
+
+    def setup_method(self) -> None:
+        os.environ["API_KEY"] = "test_key"
+        os.environ["FEATURE_PREMIUM_NUTRITION"] = "true"
+        self.client = TestClient(cast(ASGIApp, fastapi_app))
+
+    def test_premium_bmr_unavailable(self):
+        """Test premium BMR endpoint when nutrition module unavailable."""
+        with (
+            patch.dict(os.environ, {"API_KEY": "test_key"}),
+            patch("legacy_app.calculate_all_bmr", None),
+            patch("legacy_app.calculate_all_tdee", None),
+        ):
+            headers = {"X-API-Key": "test_key"}
+            data = {
+                "weight_kg": 70.0,
+                "height_cm": 175.0,
+                "age": 30,
+                "sex": "male",
+                "activity": "moderate",
+            }
+
+            response = self.client.post("/api/v1/premium/bmr", json=data, headers=headers)
+            # The endpoint actually works correctly and returns 200
+            assert response.status_code == 200
+
+    def test_premium_bmr_runtime_patch_returns_stub_response(self) -> None:
+        """Cover the conservative BMR/TDEE fallback when runtime exports are patched away."""
+        with (
+            patch.dict(os.environ, {"API_KEY": "test_key"}),
+            patch("app.calculate_all_bmr", None),
+            patch("app.calculate_all_tdee", None),
+            patch("legacy_app.calculate_all_bmr", None),
+            patch("legacy_app.calculate_all_tdee", None),
+        ):
+            headers = {"X-API-Key": "test_key"}
+            data = {
+                "weight_kg": 80.0,
+                "height_cm": 180.0,
+                "age": 35,
+                "sex": "male",
+                "activity": "light",
+                "lang": "en",
+            }
+
+            response = self.client.post("/api/v1/premium/bmr", json=data, headers=headers)
+
+        assert response.status_code == 200
+        assert response.headers.get("content-type", "").startswith("application/json")
+        body = response.json()
+        assert body["bmr"] == {"stub": 1920.0}
+        assert body["tdee"] == {"stub": 2640.0}
+        assert body["activity_level"] == "Light activity"
+        assert body["recommended_intake"]["weight_loss"] == 2112.0
+        assert body["recommended_intake"]["weight_gain"] == 3168.0
+        assert body["formulas_used"] == ["stub"]
+
+
+class TestLegacyBMRAppHelper:
+    """Legacy wrapper integration formerly owned by the app helper suite."""
+
+    def setup_method(self) -> None:
+        auth_env_name = "API_KEY"
+        os.environ[auth_env_name] = "test-key"
+        self.client = get_client()
+
+    def teardown_method(self) -> None:
+        if "API_KEY" in os.environ:
+            del os.environ["API_KEY"]
+
+    def test_bmr_wrapper_and_tdee_wrapper(self):
+        # BMR wrapper should produce a dict of numeric values
+        bmr = app_package._calculate_all_bmr_wrapper(70, 175, 30, "male", bodyfat=15)
+        assert isinstance(bmr, dict) and bmr
+        # TDEE wrapper should map BMR dict to TDEE dict with same keys
+        tdee = app_package._calculate_all_tdee_wrapper(bmr, "moderate")
+        assert isinstance(tdee, dict) and set(tdee.keys()) == set(bmr.keys())
+        assert all(isinstance(v, (int, float)) for v in tdee.values())
+
+
+class TestLegacyBMRHandlerExceptions:
+    """Tests for app.py lines 3304-3315 (premium_bmr legacy endpoint exceptions)."""
+
+    def test_premium_bmr_legacy_import_error(self, client: TestClient) -> None:
+        """/premium_bmr returns 503 when BMR calculation module unavailable."""
+
+        # Patch wrapper to raise ImportError
+        with patch.object(
+            app_package,
+            "_calculate_all_bmr_wrapper",
+            side_effect=ImportError("Module missing"),
+        ):
+            response = client.post(
+                "/premium_bmr",
+                json={
+                    "weight_kg": 70.0,
+                    "height_cm": 170.0,
+                    "age": 30,
+                    "sex": "male",
+                    "activity": "moderate",
+                    "lang": "en",
+                },
+            )
+            assert response.status_code == 503
+            data = response.json()
+            assert "BMR calculation module not available" in data["detail"]
+
+    def test_premium_bmr_legacy_value_error(self, client: TestClient) -> None:
+        """/premium_bmr returns 400 for invalid input values."""
+
+        # Patch wrapper to raise ValueError
+        with patch.object(
+            app_package,
+            "_calculate_all_bmr_wrapper",
+            side_effect=ValueError("Invalid weight"),
+        ):
+            response = client.post(
+                "/premium_bmr",
+                json={
+                    "weight_kg": 70.0,
+                    "height_cm": 170.0,
+                    "age": 30,
+                    "sex": "male",
+                    "activity": "moderate",
+                    "lang": "en",
+                },
+            )
+            assert response.status_code == 400
+            data = response.json()
+            assert "Invalid input" in data["detail"]
+            assert "Invalid weight" in data["detail"]
+
+    def test_premium_bmr_legacy_generic_exception(self, client: TestClient) -> None:
+        """/premium_bmr returns 500 for unexpected errors."""
+
+        # Patch wrapper to raise generic exception
+        with patch.object(
+            app_package,
+            "_calculate_all_bmr_wrapper",
+            side_effect=RuntimeError("Unexpected error"),
+        ):
+            response = client.post(
+                "/premium_bmr",
+                json={
+                    "weight_kg": 70.0,
+                    "height_cm": 170.0,
+                    "age": 30,
+                    "sex": "male",
+                    "activity": "moderate",
+                    "lang": "en",
+                },
+            )
+            assert response.status_code == 500
+            data = response.json()
+            assert "BMR calculation failed" in data["detail"]
+
+    def test_premium_bmr_legacy_success(self, client: TestClient) -> None:
+        """/premium_bmr returns valid response for correct inputs.
+
+        This is an integration test that verifies the actual implementation
+        produces valid values, providing valuable coverage beyond mocked tests.
+        """
+        response = client.post(
+            "/premium_bmr",
+            json={
+                "weight_kg": 70.0,
+                "height_cm": 170.0,
+                "age": 30,
+                "sex": "male",
+                "activity": "moderate",
+                "lang": "en",
+            },
+        )
+        assert response.status_code == 200
+        data = response.json()
+        assert "bmr" in data
+        assert "tdee" in data
+        assert "recommended_intake" in data
+
+        # Basic sanity checks on values
+        # BMR and TDEE are dicts with formula names as keys (e.g., {"mifflin": 1617.5})
+        assert isinstance(data["bmr"], dict), "BMR should be a dict"
+        assert isinstance(data["tdee"], dict), "TDEE should be a dict"
+        assert len(data["bmr"]) > 0, "BMR dict should not be empty"
+        assert len(data["tdee"]) > 0, "TDEE dict should not be empty"
+
+        # Check that BMR values are positive
+        for formula, value in data["bmr"].items():
+            assert value > 0, f"BMR[{formula}] should be positive"
+
+        # Check that TDEE values exceed corresponding BMR values
+        for formula in data["bmr"]:
+            if formula in data["tdee"]:
+                assert (
+                    data["tdee"][formula] > data["bmr"][formula]
+                ), f"TDEE[{formula}] should exceed BMR[{formula}]"
+
+        assert isinstance(data["recommended_intake"], dict), "recommended_intake should be a dict"
+
+
+class TestLegacyPremiumBMRErrorPaths:
+    """Legacy BMR endpoint error paths formerly owned by missing-lines coverage."""
+
+    def setup_method(self) -> None:
+        os.environ["API_KEY"] = "test_key"
+        self.client = get_client()
+
+    def teardown_method(self) -> None:
+        os.environ.pop("API_KEY", None)
+        client_instance = getattr(self, "client", None)
+        if client_instance is not None:
+            client_instance.close()
+
+    def test_premium_bmr_value_and_http_errors(self):
+        # Test premium BMR endpoint - ValueError should return 400 Bad Request
+        with (
+            patch.object(
+                app_package,
+                "calculate_all_bmr",
+                side_effect=ValueError("bad"),
+            ),
+            patch.object(app_package, "calculate_all_tdee", lambda *a, **k: {}),
+        ):
+            data = {
+                "weight_kg": 70.0,
+                "height_cm": 175.0,
+                "age": 30,
+                "sex": "male",
+                "activity": "light",
+                "lang": "en",
+            }
+            r = self.client.post(
+                "/api/v1/premium/bmr",
+                json=data,
+                headers={"X-API-Key": "test_key"},
+            )
+            assert r.status_code == 400
+            assert "Invalid input" in r.json().get("detail", "")
+
+        # Trigger HTTPException passthrough re-raise
+        with (
+            patch.object(
+                app_package,
+                "calculate_all_bmr",
+                side_effect=HTTPException(status_code=418, detail="teapot"),
+            ),
+            patch.object(app_package, "calculate_all_tdee", lambda *a, **k: {}),
+        ):
+            data = {
+                "weight_kg": 70.0,
+                "height_cm": 175.0,
+                "age": 30,
+                "sex": "male",
+                "activity": "light",
+                "lang": "en",
+            }
+            r = self.client.post(
+                "/api/v1/premium/bmr",
+                json=data,
+                headers={"X-API-Key": "test_key"},
+            )
+            assert r.status_code == 418
+            assert "teapot" in r.json().get("detail", "")
+
+
+class TestLegacyBMRImportFallback:
+    """Legacy import fallback formerly owned by simple coverage."""
+
+    _client_instance: TestClient | None = None
+
+    def setup_method(self) -> None:
+        """Set up test client."""
+        self._client_instance = TestClient(cast(ASGIApp, fastapi_app))
+
+    @property
+    def app_client(self) -> TestClient:
+        """Return initialized test client."""
+        assert self._client_instance is not None
+        return self._client_instance
+
+    def test_app_import_fallbacks(self):
+        """Test app import fallbacks."""
+        with patch("legacy_app.calculate_all_bmr", None):
+            with patch("legacy_app.calculate_all_tdee", None):
+                response = self.app_client.get("/")
+                assert response.status_code == 200
+
+
+def test_premium_bmr_resolve_wrapper_prefers_patched_app(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    async def _run() -> None:
+        """Cover api_premium_bmr wrapper resolution that returns patched callable from app package."""
+        monkeypatch.setenv("FEATURE_PREMIUM_NUTRITION", "true")
+
+        # Patch wrappers on app package so api_premium_bmr picks them up via sys.modules["app"].
+        def bmr_wrapper(*_a: Any, **_kw: Any) -> dict[str, float]:
+            return {"mifflin": 1000.0}
+
+        def tdee_wrapper(*_a: Any, **_kw: Any) -> dict[str, float]:
+            return {"mifflin": 2000.0}
+
+        monkeypatch.setattr(
+            app_package,
+            "_calculate_all_bmr_wrapper",
+            bmr_wrapper,
+            raising=False,
+        )
+        monkeypatch.setattr(
+            app_package,
+            "_calculate_all_tdee_wrapper",
+            tdee_wrapper,
+            raising=False,
+        )
+
+        req = legacy_app.BMRRequest(
+            weight_kg=70.0,
+            height_cm=175.0,
+            age=30,
+            sex="male",
+            activity="moderate",
+            bodyfat=None,
+            lang="en",
+        )
+        resp = await legacy_app.api_premium_bmr(req)
+        assert resp.bmr
+
+    asyncio.run(_run())
+
+
+def test_premium_bmr_resolve_wrapper_uses_pkg_candidates(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    async def _run() -> None:
+        """Cover api_premium_bmr wrapper resolution that returns a candidate from _iter_app_modules."""
+        monkeypatch.setenv("FEATURE_PREMIUM_NUTRITION", "true")
+
+        dummy_mod = ModuleType("dummy_app_module")
+
+        def bmr_wrapper(*_a: Any, **_kw: Any) -> dict[str, float]:
+            return {"mifflin": 1100.0}
+
+        def tdee_wrapper(*_a: Any, **_kw: Any) -> dict[str, float]:
+            return {"mifflin": 2100.0}
+
+        setattr(dummy_mod, "_calculate_all_bmr_wrapper", bmr_wrapper)
+        setattr(dummy_mod, "_calculate_all_tdee_wrapper", tdee_wrapper)
+
+        # Ensure sys.modules["app"] doesn't short-circuit the resolution
+        # app is a PEP 562 forwarding module; delattr() would trigger __getattr__ and fail even when
+        # the attribute is not actually present on the module. Remove only real module attributes.
+        monkeypatch.delitem(
+            app_package.__dict__,
+            "_calculate_all_bmr_wrapper",
+            raising=False,
+        )
+        monkeypatch.delitem(
+            app_package.__dict__,
+            "_calculate_all_tdee_wrapper",
+            raising=False,
+        )
+
+        monkeypatch.setattr(legacy_app, "_iter_app_modules", lambda: [dummy_mod])
+
+        req = legacy_app.BMRRequest(
+            weight_kg=70.0,
+            height_cm=175.0,
+            age=30,
+            sex="male",
+            activity="moderate",
+            bodyfat=None,
+            lang="en",
+        )
+        resp = await legacy_app.api_premium_bmr(req)
+        assert resp.tdee
+
+    asyncio.run(_run())
+
+
+def test_premium_bmr_legacy_executes_wrapper_resolution(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    async def _run() -> None:
+        """Cover premium_bmr_legacy wrapper resolution return path."""
+        monkeypatch.setenv("FEATURE_PREMIUM_NUTRITION", "true")
+
+        def bmr_wrapper(*_a: Any, **_kw: Any) -> dict[str, float]:
+            return {"mifflin": 1000.0}
+
+        def tdee_wrapper(*_a: Any, **_kw: Any) -> dict[str, float]:
+            return {"mifflin": 2000.0}
+
+        monkeypatch.setattr(
+            app_package,
+            "_calculate_all_bmr_wrapper",
+            bmr_wrapper,
+            raising=False,
+        )
+        monkeypatch.setattr(
+            app_package,
+            "_calculate_all_tdee_wrapper",
+            tdee_wrapper,
+            raising=False,
+        )
+
+        req = legacy_app.BMRRequestLegacy(
+            weight_kg=70.0,
+            height_cm=175.0,
+            age=30,
+            sex="male",
+            activity="moderate",
+            bodyfat=None,
+            lang="en",
+        )
+        resp = await legacy_app.premium_bmr_legacy(req)
+        assert resp.bmr
+
+    asyncio.run(_run())
+
+
+def test_premium_bmr_legacy_hits_globals_fallback_path(
+    monkeypatch: pytest.MonkeyPatch,
+) -> None:
+    async def _run() -> None:
+        """Cover premium_bmr_legacy _resolve_wrapper final globals() return (line ~4007)."""
+        monkeypatch.setenv("FEATURE_PREMIUM_NUTRITION", "true")
+
+        # app is a PEP 562 forwarding module; delattr() would trigger __getattr__ and fail even when
+        # the attribute is not actually present on the module. Remove only real module attributes.
+        monkeypatch.delitem(
+            app_package.__dict__,
+            "_calculate_all_bmr_wrapper",
+            raising=False,
+        )
+        monkeypatch.delitem(
+            app_package.__dict__,
+            "_calculate_all_tdee_wrapper",
+            raising=False,
+        )
+
+        monkeypatch.setattr(
+            legacy_app,
+            "_calculate_all_bmr_wrapper",
+            lambda *_a, **_k: {"mifflin": 1000.0},
+        )
+        monkeypatch.setattr(
+            legacy_app,
+            "_calculate_all_tdee_wrapper",
+            lambda *_a, **_k: {"mifflin": 2000.0},
+        )
+
+        req = legacy_app.BMRRequestLegacy(
+            weight_kg=70.0,
+            height_cm=175.0,
+            age=30,
+            sex="male",
+            activity="moderate",
+            bodyfat=None,
+            lang="en",
+        )
+        resp = await legacy_app.premium_bmr_legacy(req)
+        assert resp.bmr
+
+    asyncio.run(_run())
 
 
 if __name__ == "__main__":
